+++ src/main/resources/db/changelog/changes/009-drop-allegro-invoice-sequence.sql
@@ -1,0 +1,4 @@
+--liquibase formatted sql
+
+--changeset ksiegowosc:009-drop-allegro-invoice-sequence
+DROP TABLE allegro_invoice_sequence;